+++ src/SmartComponents/SystemAdvisor/SystemAdvisor.js
@@ -1,21 +1,10 @@
 import './SystemAdvisor.scss';
 
-import { AnsibeTowerIcon } from '@patternfly/react-icons';
-import {
-  BASE_URL,
-  FILTER_CATEGORIES as FC,
-  IMPACT_LABEL,
-  LIKELIHOOD_LABEL,
-  RULE_CATEGORIES,
-} from '../../AppConstants';
-import {
-  Card,
-  CardBody,
-  Tooltip,
-  TooltipPosition,
-} from '@patternfly/react-core';
+import { BASE_URL, FILTER_CATEGORIES as FC } from '../../AppConstants';
+import { Card, CardBody } from '@patternfly/react-core';
 import { useIntl } from 'react-intl';
 import React, { Fragment, useEffect, useRef, useState } from 'react';
+
 import {
   SortByDirection,
   Table,
@@ -25,29 +14,26 @@ import {
   fitContent,
   sortable,
 } from '@patternfly/react-table';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 
-import DateFormat from '@redhat-cloud-services/frontend-components/DateFormat';
 import { Get } from '../../Utilities/Api';
-import InsightsLabel from '@redhat-cloud-services/frontend-components/InsightsLabel';
 import { List } from 'react-content-loader';
 import PrimaryToolbar from '@redhat-cloud-services/frontend-components/PrimaryToolbar';
 import PropTypes from 'prop-types';
 import RemediationButton from '@redhat-cloud-services/frontend-components-remediations/RemediationButton';
-import { ReportDetails } from '@redhat-cloud-services/frontend-components-advisor-components/ReportDetails';
-import RuleLabels from '../../PresentationalComponents/Labels/RuleLabels';
+
 import { addNotification as addNotificationAction } from '@redhat-cloud-services/frontend-components-notifications/';
-import { capitalize } from '../../PresentationalComponents/Common/Tables';
 import messages from '../../Messages';
-import {
-  NoMatchingRecommendations,
-  NoRecommendations,
-  InsightsNotEnabled,
-  InventoryReportFetchFailed,
-} from './EmptyStates';
+
 import NotConnected from '@redhat-cloud-services/frontend-components/NotConnected';
-import { useLocation } from 'react-router-dom';
 import get from 'lodash/get';
+import {
+  processRemediation,
+  buildFilterChips,
+  buildRows,
+  activeRuleFirst,
+} from './helper';
+import { useDispatch } from 'react-redux';
 
 const BaseSystemAdvisor = ({ entity }) => {
   const intl = useIntl();
@@ -102,498 +88,6 @@ const BaseSystemAdvisor = ({ entity }) => {
     },
   ];
 
-  const onExpandAllClick = (_e, isOpen) => {
-    setIsAllExpanded(isOpen);
-    const allRows = [...rows];
-
-    allRows.map((row) => {
-      if (Object.prototype.hasOwnProperty.call(row, 'isOpen')) {
-        row.isOpen = isOpen;
-      }
-    });
-
-    setRows(allRows);
-  };
-
-  const onRemediationCreated = (result) => {
-    onBulkSelect(false);
-    try {
-      result.remediation && addNotification(result.getNotification());
-    } catch (error) {
-      addNotification({
-        variant: 'danger',
-        dismissable: true,
-        title: intl.formatMessage(messages.error),
-        description: `${error}`,
-      });
-    }
-  };
-
-  const actions = [
-    <RemediationButton
-      key="remediation-button"
-      isDisabled={selectedAnsibleRules.length === 0}
-      dataProvider={() => processRemediation(selectedAnsibleRules)}
-      onRemediationCreated={(result) => onRemediationCreated(result)}
-    >
-      {intl.formatMessage(messages.remediate)}
-    </RemediationButton>,
-  ];
-
-  const activeRuleFirst = (activeReports) => {
-    const reports = [...activeReports];
-    const activeRuleIndex =
-      routerData && typeof routerData.params !== 'undefined'
-        ? activeReports.findIndex(
-            (report) => report.rule.rule_id === routerData.params.id
-          )
-        : -1;
-    const activeReport = reports.splice(activeRuleIndex, 1);
-
-    return activeRuleIndex !== -1
-      ? [activeReport[0], ...reports]
-      : activeReports;
-  };
-
-  const handleOnCollapse = (_e, rowId, isOpen) => {
-    const collapseRows = [...rows];
-    collapseRows[rowId] = { ...collapseRows[rowId], isOpen };
-    setRows(collapseRows);
-  };
-  const location = useLocation().pathname?.split('/');
-
-  const buildRows = (
-    activeReports,
-    kbaDetails,
-    filters,
-    rows,
-    searchValue = '',
-    kbaLoading = false,
-    isFirstLoad = false
-  ) => {
-    const url = window.location.href;
-    let newActiveReportsList = activeReports;
-    let isRulePresent = url.indexOf('activeRule') > -1 ? true : false;
-    if (isRulePresent && isFirstLoad) {
-      let activeRule = location[2];
-      //sorts activeReportsList by making the activeRecommendation ruleId having a higher priority when sorting, or by total_risk
-      newActiveReportsList.sort((x, y) =>
-        x.rule.rule_id === activeRule
-          ? -1
-          : y.rule.rule_id === activeRule
-          ? 1
-          : 0
-      );
-    } else if (isFirstLoad) {
-      newActiveReportsList.sort((x, y) =>
-        x.rule.total_risk > y.rule.total_risk
-          ? -1
-          : y.rule.total_risk > x.rule.total_risk
-          ? 1
-          : 0
-      );
-    }
-
-    const builtRows = newActiveReportsList.flatMap((value, key) => {
-      const rule = value.rule;
-      const resolution = value.resolution;
-      const kbaDetail = Object.keys(kbaDetails).length
-        ? kbaDetails.filter((article) => article.id === value.rule.node_id)[0]
-        : {};
-      const match = rows.find((row) => row?.rule?.rule_id === rule.rule_id);
-      const selected = match?.selected;
-      const isOpen =
-        match?.isOpen || (isRulePresent && isFirstLoad && key === 0);
-
-      const reportRow = [
-        {
-          rule,
-          resolution,
-          //make arrow button disappear when there is no resolution
-          isOpen: resolution ? isOpen : undefined,
-          selected,
-          disableSelection: resolution ? !resolution.has_playbook : true,
-          cells: [
-            {
-              title: (
-                <span>
-                  {rule.description} <RuleLabels rule={rule} />
-                </span>
-              ),
-            },
-            {
-              title: (
-                <span>
-                  <DateFormat
-                    date={rule.publish_date}
-                    type="relative"
-                    tooltipProps={{ position: TooltipPosition.bottom }}
-                  />
-                </span>
-              ),
-            },
-            {
-              title: (
-                <div key={key}>
-                  <DateFormat
-                    date={value.impacted_date}
-                    type="relative"
-                    tooltipProps={{ position: TooltipPosition.bottom }}
-                  />
-                </div>
-              ),
-            },
-            {
-              title: (
-                <div key={key} style={{ verticalAlign: 'top' }}>
-                  <Tooltip
-                    key={key}
-                    position={TooltipPosition.bottom}
-                    content={
-                      <span>
-                        The <strong>likelihood</strong> that this will be a
-                        problem is {LIKELIHOOD_LABEL[rule.likelihood]}. The{' '}
-                        <strong>impact</strong> of the problem would be &nbsp;
-                        {IMPACT_LABEL[rule.impact.impact]} if it occurred.
-                      </span>
-                    }
-                  >
-                    <InsightsLabel value={rule.total_risk} isCompact />
-                  </Tooltip>
-                </div>
-              ),
-            },
-            {
-              title: (
-                <div className="ins-c-center-text" key={key}>
-                  {resolution === null ? (
-                    intl.formatMessage(messages.notAvailable)
-                  ) : resolution?.has_playbook ? (
-                    <span>
-                      <AnsibeTowerIcon size="sm" />{' '}
-                      {intl.formatMessage(messages.playbook)}
-                    </span>
-                  ) : (
-                    intl.formatMessage(messages.manual)
-                  )}
-                </div>
-              ),
-            },
-          ],
-        },
-        resolution && {
-          parent: key,
-          fullWidth: true,
-          cells: [
-            {
-              title: (
-                <ReportDetails
-                  key={`child-${key}`}
-                  report={{
-                    ...value,
-                    resolution: value.resolution.resolution,
-                  }}
-                  kbaDetail={kbaDetail}
-                  kbaLoading={kbaLoading}
-                />
-              ),
-            },
-          ],
-        },
-      ];
-      const isValidSearchValue =
-        searchValue.length === 0 ||
-        rule.description.toLowerCase().includes(searchValue.toLowerCase());
-      const isValidFilterValue =
-        Object.keys(filters).length === 0 ||
-        Object.keys(filters)
-          .map((key) => {
-            const filterValues = filters[key];
-            const rowValue = {
-              has_playbook: value.resolution?.has_playbook,
-              publish_date: rule.publish_date,
-              total_risk: rule.total_risk,
-              category: RULE_CATEGORIES[rule.category.name.toLowerCase()],
-            };
-            return filterValues.find(
-              (value) => String(value) === String(rowValue[key])
-            );
-          })
-          .every((x) => x);
-
-      return isValidSearchValue && isValidFilterValue
-        ? reportRow.filter((row) => row !== null)
-        : [];
-    });
-    //must recalculate parent for expandable table content whenever the array size changes
-    builtRows.forEach((row, index) =>
-      row.parent ? (row.parent = index - 1) : null
-    );
-
-    systemAdvisorRef.current.rowCount = activeReports.length;
-
-    if (activeReports.length < 1 || builtRows.length < 1) {
-      let EmptyState =
-        (builtRows.length === 0 && NoMatchingRecommendations) ||
-        (entity.insights_id && NoRecommendations) ||
-        InsightsNotEnabled;
-
-      return [
-        {
-          heightAuto: true,
-          cells: [
-            {
-              props: { colSpan: 5 },
-              title: <EmptyState />,
-            },
-          ],
-        },
-      ];
-    }
-
-    if (inventoryReportFetchStatus === 'failed') {
-      return [
-        {
-          heightAuto: true,
-          cells: [
-            {
-              props: { colSpan: 5 },
-              title: <InventoryReportFetchFailed entity={entity} />,
-            },
-          ],
-        },
-      ];
-    }
-
-    return builtRows;
-  };
-
-  const onRowSelect = (_e, isSelected, rowId) =>
-    setRows(
-      buildRows(
-        activeReports,
-        kbaDetailsData,
-        filters,
-        rows.map((row, index) =>
-          index === rowId ? { ...row, selected: isSelected } : row
-        ),
-        searchValue
-      )
-    );
-
-  const onBulkSelect = (isSelected) => {
-    setIsSelected(isSelected);
-    setRows(
-      buildRows(
-        activeReports,
-        kbaDetailsData,
-        filters,
-        rows.map((row, index) =>
-          // We need to use mod 2 here to ignore children with no has_playbook param
-          index % 2 === 0 && row.resolution.has_playbook
-            ? { ...row, selected: isSelected }
-            : row
-        ),
-        searchValue
-      )
-    );
-  };
-  const checkedStatus = () => {
-    if (selectedItemsLength === systemAdvisorRef.current.rowCount) {
-      return 1;
-    } else if (
-      selectedItemsLength > 0 ||
-      selectableItemsLength !== systemAdvisorRef.current.rowCount
-    ) {
-      return null;
-    } else {
-      return 0;
-    }
-  };
-
-  const bulkSelect = {
-    items: [
-      {
-        title: 'Select none',
-        onClick: () => onBulkSelect(false),
-      },
-      {
-        title: 'Select all',
-        onClick: () => onBulkSelect(true),
-      },
-    ],
-    count: selectedItemsLength,
-    checked: checkedStatus(),
-    onSelect: () => onBulkSelect(!isSelected),
-  };
-
-  const buildFilterChips = (filters) => {
-    const prunedFilters = Object.entries(filters);
-    let chips =
-      filters && prunedFilters.length > 0
-        ? prunedFilters.map((item) => {
-            const category = FC[item[0]];
-            const chips = item[1].map((value) => ({
-              name: category.values.find(
-                (values) => values.value === String(value)
-              ).label,
-              value,
-            }));
-            return {
-              category: capitalize(category.title),
-              chips,
-              urlParam: category.urlParam,
-            };
-          })
-        : [];
-    searchValue.length > 0 &&
-      chips.push({
-        category: 'Description',
-        chips: [{ name: searchValue, value: searchValue }],
-      });
-    return chips;
-  };
-
-  const onChipDelete = (_e, itemsToRemove, isAll) => {
-    if (isAll) {
-      setRows(buildRows(activeReports, kbaDetailsData, {}, rows, ''));
-      setFilters({});
-      setSearchValue('');
-    } else {
-      itemsToRemove.map((item) => {
-        if (item.category === 'Description') {
-          setRows(buildRows(activeReports, kbaDetailsData, filters, rows, ''));
-          setSearchValue('');
-        } else {
-          onFilterChange(
-            item.urlParam,
-            filters[item.urlParam].filter(
-              (value) => String(value) !== String(item.chips[0].value)
-            )
-          );
-        }
-      });
-    }
-  };
-
-  const activeFiltersConfig = {
-    deleteTitle: intl.formatMessage(messages.resetFilters),
-    filters: buildFilterChips(filters),
-    onDelete: onChipDelete,
-  };
-
-  const fetchKbaDetails = async (reportsData) => {
-    const kbaIds = reportsData.map(({ rule }) => rule.node_id).filter((x) => x);
-    try {
-      const kbaDetailsFetch = (
-        await Get(
-          `https://access.redhat.com/hydra/rest/search/kcs?q=id:(${kbaIds.join(
-            ` OR `
-          )})&fl=view_uri,id,publishedTitle&rows=${
-            kbaIds.length
-          }&redhat_client=$ADVISOR`,
-          {},
-          { credentials: 'include' }
-        )
-      ).data.response.docs;
-
-      setKbaDetailsData(kbaDetailsFetch);
-      setRows(
-        buildRows(
-          reportsData,
-          kbaDetailsFetch,
-          filters,
-          rows,
-          searchValue,
-          false,
-          true
-        )
-      );
-    } catch (error) {
-      console.error(error, 'KBA fetch failed.');
-    }
-  };
-
-  const onSort = (_e, index, direction) => {
-    const sortedReports = {
-      2: 'rule.description',
-      3: 'rule.publish_date',
-      4: 'impacted_date',
-      5: 'rule.total_risk',
-      6: 'resolution.has_playbook',
-    };
-    const d = direction === SortByDirection.asc ? 1 : -1;
-
-    const sort = () =>
-      activeReports.concat().sort((firstItem, secondItem) => {
-        let fst = get(firstItem, sortedReports[index]);
-        let snd = get(secondItem, sortedReports[index]);
-
-        if (index === 3 || index === 4) {
-          fst = new Date(fst);
-          snd = new Date(snd);
-        }
-        return fst > snd ? d : snd > fst ? -d : 0;
-      });
-
-    const sortedReportsDirectional = sort();
-
-    setActiveReports(sortedReportsDirectional);
-    setSortBy({
-      index,
-      direction,
-    });
-    setRows(
-      buildRows(
-        sortedReportsDirectional,
-        kbaDetailsData,
-        filters,
-        rows,
-        searchValue
-      )
-    );
-  };
-
-  const onFilterChange = (param, values) => {
-    const removeFilterParam = (param) => {
-      const filter = { ...filters };
-      delete filter[param];
-      return filter;
-    };
-
-    const newFilters =
-      values.length > 0
-        ? { ...filters, ...{ [param]: values } }
-        : removeFilterParam(param);
-    setRows(
-      buildRows(activeReports, kbaDetailsData, newFilters, rows, searchValue)
-    );
-    setFilters(newFilters);
-  };
-
-  const onInputChange = (value) => {
-    const builtRows = buildRows(
-      activeReports,
-      kbaDetailsData,
-      filters,
-      rows,
-      value
-    );
-    setSearchValue(value);
-    setRows(builtRows);
-  };
-
-  const processRemediation = (selectedAnsibleRules) => {
-    const playbookRows = selectedAnsibleRules.filter(
-      (r) => r.resolution?.has_playbook
-    );
-    const issues = playbookRows.map((r) => ({
-      id: `advisor:${r.rule.rule_id}`,
-      description: r.rule.description,
-    }));
-    return issues.length ? { issues, systems: [entity.id] } : false;
-  };
-
   const filterConfigItems = [
     {
       label: 'description',
@@ -643,6 +137,307 @@ const BaseSystemAdvisor = ({ entity }) => {
     },
   ];
 
+  const actions = [
+    <RemediationButton
+      key="remediation-button"
+      isDisabled={selectedAnsibleRules.length === 0}
+      dataProvider={() => processRemediation(selectedAnsibleRules, entity)}
+      onRemediationCreated={(result) =>
+        onRemediationCreated(result, onBulkSelect)
+      }
+    >
+      {intl.formatMessage(messages.remediate)}
+    </RemediationButton>,
+  ];
+
+  const onExpandAllClick = (_e, isOpen) => {
+    setIsAllExpanded(isOpen);
+    const allRows = [...rows];
+
+    allRows.map((row) => {
+      if (Object.prototype.hasOwnProperty.call(row, 'isOpen')) {
+        row.isOpen = isOpen;
+      }
+    });
+
+    setRows(allRows);
+  };
+
+  const handleOnCollapse = (_e, rowId, isOpen) => {
+    const collapseRows = [...rows];
+    collapseRows[rowId] = { ...collapseRows[rowId], isOpen };
+    setRows(collapseRows);
+  };
+
+  const onRowSelect = (_e, isSelected, rowId) =>
+    setRows(
+      buildRows(
+        activeReports,
+        kbaDetailsData,
+        filters,
+        rows.map((row, index) =>
+          index === rowId ? { ...row, selected: isSelected } : row
+        ),
+        searchValue,
+        false,
+        false,
+        location,
+        systemAdvisorRef,
+        entity,
+        intl
+      )
+    );
+
+  const onBulkSelect = (isSelected) => {
+    setIsSelected(isSelected);
+    setRows(
+      buildRows(
+        activeReports,
+        kbaDetailsData,
+        filters,
+        rows.map((row, index) =>
+          // We need to use mod 2 here to ignore children with no has_playbook param
+          index % 2 === 0 && row.resolution.has_playbook
+            ? { ...row, selected: isSelected }
+            : row
+        ),
+        searchValue,
+        false,
+        false,
+        location,
+        systemAdvisorRef,
+        entity,
+        intl
+      )
+    );
+  };
+  const checkedStatus = () => {
+    if (selectedItemsLength === systemAdvisorRef.current.rowCount) {
+      return 1;
+    } else if (
+      selectedItemsLength > 0 ||
+      selectableItemsLength !== systemAdvisorRef.current.rowCount
+    ) {
+      return null;
+    } else {
+      return 0;
+    }
+  };
+
+  const bulkSelect = {
+    items: [
+      {
+        title: 'Select none',
+        onClick: () => onBulkSelect(false),
+      },
+      {
+        title: 'Select all',
+        onClick: () => onBulkSelect(true),
+      },
+    ],
+    count: selectedItemsLength,
+    checked: checkedStatus(),
+    onSelect: () => onBulkSelect(!isSelected),
+  };
+
+  const onChipDelete = (_e, itemsToRemove, isAll) => {
+    if (isAll) {
+      setRows(
+        buildRows(
+          activeReports,
+          kbaDetailsData,
+          {},
+          rows,
+          '',
+          false,
+          false,
+          location,
+          systemAdvisorRef,
+          entity,
+          intl
+        )
+      );
+      setFilters({});
+      setSearchValue('');
+    } else {
+      itemsToRemove.map((item) => {
+        if (item.category === 'Description') {
+          setRows(
+            buildRows(
+              activeReports,
+              kbaDetailsData,
+              filters,
+              rows,
+              '',
+              false,
+              false,
+              location,
+              systemAdvisorRef,
+              entity,
+              intl
+            )
+          );
+          setSearchValue('');
+        } else {
+          onFilterChange(
+            item.urlParam,
+            filters[item.urlParam].filter(
+              (value) => String(value) !== String(item.chips[0].value)
+            )
+          );
+        }
+      });
+    }
+  };
+  const onRemediationCreated = (result, onBulkSelect) => {
+    onBulkSelect(false);
+    try {
+      result.remediation && addNotification(result.getNotification());
+    } catch (error) {
+      addNotification({
+        variant: 'danger',
+        dismissable: true,
+        title: intl.formatMessage(messages.error),
+        description: `${error}`,
+      });
+    }
+  };
+
+  const activeFiltersConfig = {
+    deleteTitle: intl.formatMessage(messages.resetFilters),
+    filters: buildFilterChips(filters, searchValue),
+    onDelete: onChipDelete,
+  };
+
+  const fetchKbaDetails = async (reportsData) => {
+    const kbaIds = reportsData.map(({ rule }) => rule.node_id).filter((x) => x);
+    try {
+      const kbaDetailsFetch = (
+        await Get(
+          `https://access.redhat.com/hydra/rest/search/kcs?q=id:(${kbaIds.join(
+            ` OR `
+          )})&fl=view_uri,id,publishedTitle&rows=${
+            kbaIds.length
+          }&redhat_client=$ADVISOR`,
+          {},
+          { credentials: 'include' }
+        )
+      ).data.response.docs;
+
+      setKbaDetailsData(kbaDetailsFetch);
+      setRows(
+        buildRows(
+          reportsData,
+          kbaDetailsFetch,
+          filters,
+          rows,
+          searchValue,
+          false,
+          true,
+          location,
+          systemAdvisorRef,
+          entity,
+          intl
+        )
+      );
+    } catch (error) {
+      console.error(error, 'KBA fetch failed.');
+    }
+  };
+
+  const onSort = (_e, index, direction) => {
+    const sortedReports = {
+      2: 'rule.description',
+      3: 'rule.publish_date',
+      4: 'impacted_date',
+      5: 'rule.total_risk',
+      6: 'resolution.has_playbook',
+    };
+    const d = direction === SortByDirection.asc ? 1 : -1;
+
+    const sort = () =>
+      activeReports.concat().sort((firstItem, secondItem) => {
+        let fst = get(firstItem, sortedReports[index]);
+        let snd = get(secondItem, sortedReports[index]);
+
+        if (index === 3 || index === 4) {
+          fst = new Date(fst);
+          snd = new Date(snd);
+        }
+        return fst > snd ? d : snd > fst ? -d : 0;
+      });
+
+    const sortedReportsDirectional = sort();
+
+    setActiveReports(sortedReportsDirectional);
+    setSortBy({
+      index,
+      direction,
+    });
+    setRows(
+      buildRows(
+        sortedReportsDirectional,
+        kbaDetailsData,
+        filters,
+        rows,
+        searchValue,
+        false,
+        false,
+        location,
+        systemAdvisorRef,
+        entity,
+        intl
+      )
+    );
+  };
+
+  const onFilterChange = (param, values) => {
+    const removeFilterParam = (param) => {
+      const filter = { ...filters };
+      delete filter[param];
+      return filter;
+    };
+
+    const newFilters =
+      values.length > 0
+        ? { ...filters, ...{ [param]: values } }
+        : removeFilterParam(param);
+    setRows(
+      buildRows(
+        activeReports,
+        kbaDetailsData,
+        newFilters,
+        rows,
+        searchValue,
+        false,
+        false,
+        location,
+        systemAdvisorRef,
+        entity,
+        intl
+      )
+    );
+    setFilters(newFilters);
+  };
+
+  const onInputChange = (value) => {
+    const builtRows = buildRows(
+      activeReports,
+      kbaDetailsData,
+      filters,
+      rows,
+      value,
+      false,
+      false,
+      location,
+      systemAdvisorRef,
+      entity,
+      intl
+    );
+    setSearchValue(value);
+    setRows(builtRows);
+  };
+
   useEffect(() => {
     const dataFetch = async () => {
       try {
@@ -652,21 +447,28 @@ const BaseSystemAdvisor = ({ entity }) => {
             credentials: 'include',
           }
         );
-
-        const activeRuleFirstReportsData = activeRuleFirst(reportsFetch.data);
+        const activeRuleFirstReportsData = activeRuleFirst(
+          reportsFetch,
+          routerData
+        );
         fetchKbaDetails(activeRuleFirstReportsData);
-
         setRows(
           buildRows(
-            activeRuleFirstReportsData,
+            reportsFetch,
             {},
             filters,
             rows,
             searchValue,
-            true
+            true,
+            false,
+            location,
+            systemAdvisorRef,
+            entity,
+            intl
           )
         );
         setInventoryReportFetchStatus('fulfilled');
+        //reports fetch
         setActiveReports(activeRuleFirstReportsData);
       } catch (error) {
         setInventoryReportFetchStatus('failed');
